problem with replacing pixel values
turns out that it was the problem with JPG pictures. Not the algorithm's fault. when values are put into pixels by using putpixels() and then display the pixels, the displayed values are not the ones that have been put with the putpixels()! but this does not happen with PNG pictures.