changed algorithm for finding the closest match
incorporated method to find how far RGB value is from original RGB value.
put weight of 1.2 or 1.3 times more on how far calculation than the total difference. This means that total difference + ( how far red * 1.3 ) + ( how far green * 1.3 ) + (how far blue * 1.3 )

I also expanded the color groups so that the new picture better represents the real image The color group is the web safe colors which is about 216 colors. This is probably enough to represent the real image and still makes it far far easier to deal with! when compared to millions of combinations of pixel values!